refactor(hir): split definition and type query groups
