+++ crates/hir/src/lib.rs
@@ -8,11 +8,13 @@ pub mod def_id;
 pub mod display;
 pub mod has_source;
 pub mod hir_def;
+pub mod pathres;
 pub mod region_tree;
 pub mod scope;
 pub mod semantics;
 pub mod source_map;
 pub mod symbol;
+mod ty_db;
 pub mod type_infer;
 
 #[cfg(test)]
