+++ crates/ide/src/code_action/handlers/add_missing_connections.rs
@@ -1,5 +1,5 @@
 use hir::{
-    base_db::source_db::SourceDb, container::InModule, db::HirDb,
+    base_db::source_db::SourceDb, container::InModule, db::HirDefDb,
     hir_def::module::instantiation::PortConn, source_map::IsSrc,
 };
 use rustc_hash::FxHashSet;
