+++ crates/ide/src/code_lens.rs
@@ -1,5 +1,5 @@
 use hir::{
-    db::HirDb,
+    db::HirDefDb,
     def_id::DefId,
     file::HirFileId,
     hir_def::{
